+++ crates/optimism/trie/src/db/models/mod.rs
@@ -1,20 +1,25 @@
-//! MDBX tables and models for external proofs storage.
+//! MDBX implementation of [`OpProofsStorage`](crate::OpProofsStorage).
+//!
+//! This module provides a complete MDBX implementation of the
+//! [`OpProofsStorage`](crate::OpProofsStorage) trait. It uses the [`reth_db`] crate for
+//! database interactions and defines the necessary tables and models for storing trie branches,
+//! accounts, and storage leaves.
+
+mod block;
+pub use block::*;
+mod version;
+pub use version::*;
+mod storage;
+pub use storage::*;
+
 use alloy_primitives::B256;
-use reth_db_api::{
+use reth_db::{
     table::{DupSort, TableInfo},
     tables, TableSet, TableType, TableViewer,
 };
 use reth_primitives_traits::Account;
 use reth_trie::{BranchNodeCompact, StoredNibbles};
 use std::fmt;
-
-mod block;
-pub use block::*;
-mod version;
-pub use version::*;
-
-mod storage;
-pub use storage::*;
 
 tables! {
     /// Stores historical branch nodes for the account state trie.
