+++ crates/optimism/trie/src/db/models/version.rs
@@ -1,5 +1,8 @@
 use bytes::{Buf, BufMut};
-use reth_db_api::table::{Compress, Decompress};
+use reth_db::{
+    table::{Compress, Decompress},
+    DatabaseError,
+};
 use serde::{Deserialize, Serialize};
 
 /// Wrapper type for `Option<T>` that implements `Compress` and `Decompress`
@@ -41,7 +44,7 @@ impl<T: Compress> Compress for MaybeDeleted<T> {
 }
 
 impl<T: Decompress> Decompress for MaybeDeleted<T> {
-    fn decompress(value: &[u8]) -> Result<Self, reth_db_api::DatabaseError> {
+    fn decompress(value: &[u8]) -> Result<Self, DatabaseError> {
         if value.is_empty() {
             // Empty = deleted
             Ok(Self(None))
@@ -84,9 +87,9 @@ impl<T: Compress> Compress for VersionedValue<T> {
 }
 
 impl<T: Decompress> Decompress for VersionedValue<T> {
-    fn decompress(value: &[u8]) -> Result<Self, reth_db_api::DatabaseError> {
+    fn decompress(value: &[u8]) -> Result<Self, DatabaseError> {
         if value.len() < 8 {
-            return Err(reth_db_api::DatabaseError::Decode);
+            return Err(DatabaseError::Decode);
         }
 
         let mut buf: &[u8] = value;
